Number guessing game in spanish
This mini game goes on a limit of numbers from 0 to 100. 
It was made in an effort to learn more about programming in my early beginnings. 
The focus here is on structure, defining and calling variables, and the usage of if, while, and more.
This code is made through VSCode. It's a Python code.